requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
asttokens @ file:///C:/b/abs_9662ywy9fp/croot/asttokens_1743630464377/work
attrs==25.3.0
backoff==2.2.1
backports.tarfile==1.2.0
banks==2.1.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.37.34
botocore==1.37.34
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.1
click==8.1.8
colorama @ file:///C:/ci_311/colorama_1676422310965/work
coloredlogs==15.0.1
comm @ file:///C:/b/abs_67a8058udb/croot/comm_1709322909844/work
coverage==7.8.0
dataclasses-json==0.6.7
datasets==3.5.0
debugpy @ file:///C:/b/abs_bf9oo2vhxp/croot/debugpy_1736269476451/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepeval==2.5.0
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
docutils==0.21.2
docx2txt==0.9
durationpy==0.9
et_xmlfile==2.0.0
execnet==2.1.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.39.0
google-genai==1.11.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
griffe==1.7.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.7.9
ipykernel @ file:///C:/b/abs_6c9ggygp01/croot/ipykernel_1737660720620/work
ipython @ file:///C:/b/abs_8eyhzleyrk/croot/ipython_1734548134403/work
ipywidgets==8.1.6
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client @ file:///C:/b/abs_149bw133if/croot/jupyter_client_1737570986926/work
jupyter_core @ file:///C:/b/abs_beftpbuevw/croot/jupyter_core_1718818307097/work
jupyterlab_widgets==3.0.14
keyring==25.6.0
kubernetes==32.0.1
langchain==0.3.23
langchain-aws==0.2.21
langchain-community==0.3.21
langchain-core==0.3.54
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langsmith==0.3.31
llama-cloud==0.1.18
llama-cloud-services==0.6.12
llama-index==0.12.31
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.31
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.37
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline @ file:///C:/ci_311/matplotlib-inline_1676425798036/work
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.8
onnxruntime==1.21.0
openai==1.75.0
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
parso @ file:///C:/b/abs_834b4mj92b/croot/parso_1733963322289/work
pathspec==0.12.1
pillow==11.2.1
pkginfo==1.10.0
platformdirs @ file:///C:/b/abs_ddh15014or/croot/platformdirs_1744273060660/work
pluggy==1.5.0
portalocker==3.1.1
posthog==3.25.0
prompt-toolkit @ file:///C:/b/abs_68uwr58ed1/croot/prompt-toolkit_1704404394082/work
propcache==0.3.1
protobuf==5.29.4
psutil @ file:///C:/b/abs_b5gv3mn55h/croot/psutil_1736371546320/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments @ file:///C:/b/abs_e4bg5vh5j_/croot/pygments_1744667628203/work
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-repeat==0.9.4
pytest-rerunfailures==12.0
pytest-xdist==3.6.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.1.0
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq @ file:///C:/b/abs_f3yte6j5yn/croot/pyzmq_1734711069724/work
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.11.4
sentry-sdk==2.26.1
shellingham==1.5.4
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.46.2
striprtf==0.0.26
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.1
tornado @ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm==4.67.1
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
twine==5.1.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
